refactor: move zoxide add to separate file
